Factor out lowering of memory ops (#102)
